requirements.txt
coverage==6.3.2
Flask==2.1.1
pydantic==1.9.0
python-dotenv==0.20.0
typing_extensions==4.1.1
lyricsgenius==3.0.1
boto3==1.22.9
redis==4.3.0